@@ -1,25 +1,25 @@
 source 'https://rubygems.org'
-ruby "1.9.3"
+ruby "2.1.2"
 
-gem 'rails', '3.2.18'
-gem 'devise', '2.1.2'
-gem 'cancan', '1.6.8'
-gem 'simple_form', '2.0.4'
+gem 'rails', '4.2.3'
+gem 'devise'
+gem 'cancan'
+gem 'simple_form'
 gem 'jquery-rails'
 gem "validates_existence"
 gem 'mysql2'
 gem "recaptcha", :require => "recaptcha/rails"
 gem "paper_trail"
 gem 'delayed_job_active_record'
-gem 'will_paginate', '> 3.0'
+gem 'will_paginate'
 gem 'newrelic_rpm'
 gem 'validates_timeliness'
 gem 'jquery-ui-rails'
 gem 'angularjs-rails'
-gem 'foreman', '0.74'
-gem 'sass-rails',   '~> 3.2.3'
-gem 'coffee-rails', '~> 3.2.1'
-gem 'uglifier', '>= 1.0.3'
+gem 'foreman'
+gem 'sass-rails'
+gem 'coffee-rails'
+gem 'uglifier'
 gem 'rabl'
 gem 'oj'
 
@@ -30,7 +30,7 @@ end
 group :development, :test do
     gem 'thin'
     gem 'rspec-rails'
-    gem 'rspec-core', '~> 2.99.1'
+    gem 'rspec-core'
     gem 'mocha'
     gem 'factory_girl_rails'
     gem 'shoulda'
@@ -43,5 +43,5 @@ group :development, :test do
 end
 
 group :development do
-    gem 'debugger'
+    gem 'byebug'
 end
