@@ -7,7 +7,7 @@ class ConferencesController < ApplicationController
   end
 
   def create
-    @conference = Conference.new(params[:conference])
+    @conference = Conference.new(conference_params)
     authorize! :create, @conference
 
     if @conference.save
@@ -19,7 +19,7 @@ class ConferencesController < ApplicationController
   end
 
   def edit
-    @conference = Conference.includes(conference_items: [:regular_prices]).find_by_conference_year(params[:id])
+    @conference = Conference.includes(conference_items: [:regular_prices]).find_by(conference_year: params[:id])
     @new_conference = Conference.new(conference_year: (@conference.conference_year + 1), tax_rate: @conference.tax_rate)
     authorize! @conference, :update
     if params[:pricing]
@@ -33,7 +33,7 @@ class ConferencesController < ApplicationController
 
   def update
     authorize! @conference, :update
-    if @conference.update_attributes(params[:conference])
+    if @conference.update_attributes(conference_params)
       redirect_to conference_path(@conference)
     else
       if params[:pricing]
@@ -58,6 +58,18 @@ class ConferencesController < ApplicationController
   private
 
     def find_conference
-      @conference = Conference.find_by_conference_year(params[:id])
+      @conference = Conference.find_by(conference_year: params[:id])
+    end
+
+    def conference_params
+      params.require(:conference).permit(
+        # conference params
+        :conference_year, :tax_rate, :active,
+          # nested conference items
+          conference_items_attributes: [:name, :description, :multiple, :max, :visibility, :manual_price, :user_comment, :user_comment_prompt, :conference_id,
+            # nested prices
+            prices_attributes: Price::WHITE_LISTED
+          ]
+      )
     end
 end
