@@ -1,6 +1,6 @@
 class PaymentMailer < ActionMailer::Base
-  default from: CONFIG[:gmail_username]
-  
+  default from: ENV["GMAIL_USERNAME"]
+
   def payment_notification(payment)
     @user = payment.user
     @payment = payment
