@@ -1,21 +1,20 @@
 
 class ConferenceItem < ActiveRecord::Base
   has_paper_trail
-  attr_accessible :name, :description, :multiple, :max, :visibility, :prices_attributes, :manual_price, :user_comment, :user_comment_prompt, :conference_id
 
   belongs_to :conference
   has_many :prices
   has_many :line_items
   has_many :itineraries, :through => :line_items
-  has_many :regular_prices, :conditions => {:discount_key => nil}, :class_name => 'Price' #exclude all of the prices created for this item that have a discount key set
+  has_many :regular_prices, -> { where :discount_key => nil }, :class_name => 'Price' #exclude all of the prices created for this item that have a discount key set
 
   after_initialize :build_price_objects, :if => "prices.blank?" #When a new conference item is created the event planners will probably want to price these six categories
 
   accepts_nested_attributes_for :prices, allow_destroy: true
 
   scope :current_active, ->{ where("visibility=? AND conference_id='?'", true, Conference.active.id) }
-  scope :not_registered, lambda{|line_items| where("conference_items.id not in (?)", line_items.blank? ? '' : line_items.collect(&:conference_item_id))}
-  scope :not_discounted, lambda{|discounted_items| where("conference_items.id not in (?)", discounted_items.blank? ? '' : discounted_items.collect(&:id))}
+  scope :not_registered, lambda {|line_items| where("conference_items.id not in (?)", line_items.blank? ? '' : line_items.collect(&:conference_item_id))}
+  scope :not_discounted, lambda {|discounted_items| where("conference_items.id not in (?)", discounted_items.blank? ? '' : discounted_items.collect(&:id))}
   #I couldn't figure out a single sql query that would inner join prices and conference items where the itinerary discount key matches price discount key, or (if no such price exists) where price country_category + member matches same fields in user
   #So instead I find the discounted items first and then price all the conference items that aren't in discounted items
 
