@@ -9,7 +9,7 @@ Iftaconferenceapp::Application.configure do
   config.action_controller.perform_caching = true
 
   # Disable Rails's static asset server (Apache or nginx will already do this)
-  config.serve_static_assets = false
+  config.serve_static_files = false
 
   # Compress JavaScripts and CSS
   config.assets.compress = true
@@ -66,6 +66,6 @@ Iftaconferenceapp::Application.configure do
   # config.active_record.auto_explain_threshold_in_seconds = 0.5
 
   # Default URL Options
-  config.action_mailer.default_url_options = {:host => CONFIG[:mailer_host]}
+  config.action_mailer.default_url_options = {:host => ENV["MAILER_HOST"]}
 
 end
