@@ -1,5 +1,5 @@
 Recaptcha.configure do |config|
-  config.public_key  = CONFIG[:recaptcha_public_key]
-  config.private_key = CONFIG[:recaptcha_private_key]
+  config.public_key  = ENV["RECAPTCHA_PUBLIC_KEY"]
+  config.private_key = ENV["RECAPTCHA_PRIVATE_KEY"]
   #config.proxy = 'http://myproxy.com.au:8080'
 end
