@@ -4,4 +4,5 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-Iftaconferenceapp::Application.config.secret_token = CONFIG[:secret_token]
+Iftaconferenceapp::Application.config.secret_token = ENV["SECRET_TOKEN"]
+Iftaconferenceapp::Application.config.secret_key_base = ENV["SECRET_KEY_BASE"]
