@@ -1,5 +1,5 @@
-require 'development_mail_interceptor' 
-ActionMailer::Base.register_interceptor(DevelopmentMailInterceptor) unless Rails.env.production?  
+require 'development_mail_interceptor'
+ActionMailer::Base.register_interceptor(DevelopmentMailInterceptor) unless Rails.env.production?
 
 # SMTP Connection Settings
 ActionMailer::Base.delivery_method = :smtp
@@ -8,9 +8,9 @@ ActionMailer::Base.raise_delivery_errors = true
 ActionMailer::Base.smtp_settings = {
 :address              => "smtp.gmail.com",
 :port                 => 587,
-:domain               => CONFIG[:gmail_domain],
-:user_name            => CONFIG[:gmail_username],
-:password             => CONFIG[:gmail_password],
+:domain               => ENV["GMAIL_DOMAIN"],
+:user_name            => ENV["GMAIL_USERNAME"],
+:password             => ENV["GMAIL_PASSWORD"],
 :authentication       => 'plain',
 :enable_starttls_auto => true  }
-  
+
