@@ -2,7 +2,7 @@ Iftaconferenceapp::Application.routes.draw do
   root :to => 'welcome#index', :as => :root
   #user routes
   get 'users/:id/edit_password' => 'users#edit_password', :as => :edit_user_password
-  match 'users/:id/update_password' => 'users#update_password', :as => :update_user_password
+  patch 'users/:id/update_password' => 'users#update_password', :as => :update_user_password
   post 'users/:id/change_role' => 'users#change_role', :as => :change_role_user
   get 'users/edit' => 'users#edit'
 
@@ -21,8 +21,8 @@ Iftaconferenceapp::Application.routes.draw do
   devise_scope :user do
     get 'users/sign_out' => 'devise/sessions#destroy'
     put 'users/password' => 'devise/passwords#update'
-    match 'users/password' => 'devise/passwords#create' #Rails bug with post member routes
-    match 'users/password/edit' => 'devise/passwords#edit', :as => :edit_password
+    post 'users/password' => 'devise/passwords#create'
+    get 'users/password/edit' => 'devise/passwords#edit', :as => :edit_password
   end
   resources :users
   resources :itineraries do
