@@ -1,6 +1,6 @@
 class DevelopmentMailInterceptor
   def self.delivering_email(message)
     message.subject = "[#{message.to}] #{message.subject}"
-    message.to = CONFIG[:gmail_username]
+    message.to = ENV["GMAIL_USERNAME"]
   end
 end
