@@ -82,7 +82,7 @@ describe Conference do
     it 'should de-activate an old conference for a new conference' do
       c1 = Conference.create({tax_rate: 0.1, conference_year: 2013, active: true})
       c2 = Conference.create({tax_rate: 0.1, conference_year: 2014, active: true})
-      expect(Conference.find(c1).active).to be_falsey
+      expect(Conference.find(c1.id).active).to be_falsey
     end
 
     it 'should force activate a conference when there is no active conference' do
@@ -96,7 +96,7 @@ describe Conference do
     it 'should return a conference_year' do
       conference = FactoryGirl.build(:conference)
       conference.conference_year = 2014
-      expect(conference.to_param).to eql(2014)
+      expect(conference.to_param).to eql("2014")
     end
   end
 end
