@@ -4,7 +4,7 @@ describe Payment do
   context 'mark_transaction on save' do
     it 'should set transaction paid to false' do
       payment = FactoryGirl.build(:unconfirmed_payment)
-      transaction = payment.transaction
+      transaction = payment.order
       transaction.paid = true
 
       transaction.expects(:save)
@@ -14,7 +14,7 @@ describe Payment do
 
     it 'should set transaction paid to true' do
       payment = FactoryGirl.build(:confirmed_payment)
-      transaction = payment.transaction
+      transaction = payment.order
       transaction.paid = false
 
       transaction.expects(:save)
@@ -27,7 +27,7 @@ describe Payment do
   context 'clear_transaction on destroy' do
     it 'should set the transaction paid flag to false on destroy' do
       payment = FactoryGirl.build(:payment)
-      transaction = payment.transaction
+      transaction = payment.order
       transaction.paid = true
 
       transaction.expects(:save)
