@@ -26,7 +26,7 @@ describe 'SignUp' do
 
     user = assigns(:user)
     country = FactoryGirl.create(:country)
-    put "users/#{user.id}", { user: {
+    put "/users/#{user.id}", { user: {
       first_name: 'Jane',
       last_name: 'Doe',
       nametag_name: 'Jane Doe',
